Add files via upload
Awesome Restaurant is your one stop destination for all your food needs. You get to choose right from your local delicacies like Idly Vada, Pav bhaji or Samosa to exotic world class delicacies from across the world like Russian, Chinese or Japanese.
The main reason for us to choose to develop a food delivery website is that the hunger for food is never ending and we would like to try to bring out ideas which could be the better version of what is already available so that the users can get a better experience.